Updated function restart_elevated

+++ scripts/util_functions.psm1
@@ -4,58 +4,35 @@ function isadmin {
 }
 
 function restart_elevated {
-
     param(
-        $script_args,
+        $arguments,
         [switch]$kill_original,
-        [string]$current_dir = $PWD.path
+        [string]$workdir = $PWD.path
     )
 
-    if ($MyInvocation.ScriptName -eq "") {
-        throw 'Script must be saved as a .ps1 file.'
-    }
-
-    if (isadmin) {
-        Write-Host 'running as administrator: OK' -ForegroundColor DarkGray
-        return $null
-    }
+    if ($MyInvocation.ScriptName -eq "") { throw 'Script must be saved as a .ps1 file.' }
+    if (isadmin) { return $null }
 
     try {
-        Write-Warning '(!) running as administrator: NO'
-        Write-Host 'Restarting with elevated permissions'
-        sleep 3
-
         $script_fullpath = $MyInvocation.ScriptName
-
         $argline = "-noprofile -nologo -noexit"
-        $argline += " -Command cd `"$current_dir`"; `"$script_fullpath`""
+        $argline += " -Command cd `"$workdir`"; `"$script_fullpath`""
 
-        $script_args.GetEnumerator() | % {
-
-            if ($_.Value -is [boolean]) {
-                $argline += " -$($_.key) `$$($_.value)"
-            }
-
-            elseif ($_.Value -is [switch]) {
-                $argline += " -$($_.key)"
-            }
-
-            else {
-                $argline += " -$($_.key) `"$($_.value)`""
+        if ($arguments) {
+            $arguments.GetEnumerator() | % {
+                if ($_.Value -is [boolean]) { $argline += " -$($_.key) `$$($_.value)" }
+                elseif ($_.Value -is [switch]) { $argline += " -$($_.key)" }
+                else { $argline += " -$($_.key) `"$($_.value)`"" }
             }
         }
 
         $p = Start-Process "$PSHOME\powershell.exe" -Verb Runas -ArgumentList $argline -PassThru -ErrorAction 'stop'
-
-        if ($kill_original) {
-            [System.Diagnostics.Process]::GetCurrentProcess() | Stop-Process -ErrorAction Stop
-        }
-
-        Write-Host "Elevated process id: $($p.id)"
+        if ($kill_original) { [System.Diagnostics.Process]::GetCurrentProcess() | Stop-Process -ErrorAction Stop }
+        write "Elevated process id: $($p.id)"
         exit
     }
     catch {
-        Write-Error "ERROR: Failed to restart script with elevated premissions."
+        error "Failed to restart script with elevated premissions."
         throw $_
     }
 }
